requirements-local.txt
torch==1.7.0
mlflow
dvc
torchvision==0.8.1
pyyaml
Pillow==7.1.2
pre-commit==2.19.0
autoflake
